refactor(websocket): centralize STOMP client to prevent duplicate connections
